Refactor the UI controller class
