+++ .vscode/settings.json
@@ -19,6 +19,7 @@
     },
     "cSpell.words": [
         "Appliable",
+        "ARGB",
         "culori",
         "okhsl",
         "Unregisters"
